+++ lib/widgets/card-post.dart
@@ -1,5 +1,7 @@
 import 'package:dvote/dvote.dart';
+import 'package:dvote_common/lib/common.dart';
 import 'package:flutter/material.dart';
+import 'package:vocdoni/app-config.dart';
 import 'package:vocdoni/data-models/entity.dart';
 import 'package:vocdoni/lib/makers.dart';
 import 'package:eventual/eventual-builder.dart';
@@ -17,12 +19,20 @@ class CardPost extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    String headerUrl = post.image;
+    if (headerUrl.startsWith("ipfs"))
+      headerUrl =
+          processIpfsImageUrl(headerUrl, ipfsDomain: AppConfig.IPFS_DOMAIN);
+    String avatarUrl = entity.metadata.value.media.avatar;
+    if (avatarUrl.startsWith("ipfs"))
+      avatarUrl =
+          processIpfsImageUrl(avatarUrl, ipfsDomain: AppConfig.IPFS_DOMAIN);
     // Consume individual items that may rebuild only themselves
     return EventualBuilder(
       notifiers: [entity.metadata, entity.feed],
       builder: (context, _, __) => BaseCard(
           onTap: () => onPostCardTap(context, post, entity),
-          image: post.image,
+          image: headerUrl,
           imageTag: makeElementTag(entity.reference.entityId, post.id, listIdx),
           children: <Widget>[
             ListItem(
@@ -31,7 +41,7 @@ class CardPost extends StatelessWidget {
               mainTextFullWidth: true,
               secondaryText: entity
                   .metadata.value.name[entity.metadata.value.languages[0]],
-              avatarUrl: entity.metadata.value.media.avatar,
+              avatarUrl: avatarUrl,
               avatarText: entity
                   .metadata.value.name[entity.metadata.value.languages[0]],
               avatarHexSource: entity.reference.entityId,